Filename too long
Issue
    jenkinsci/cucumber-reports-plugin/issues/101
    damianszczepanik/maven-cucumber-reporting/issues/53
