+++ src/test/java/net/masterthought/cucumber/generators/integrations/FeatureReportPageIntegrationTest.java
@@ -94,7 +94,7 @@ public class FeatureReportPageIntegrationTest extends PageTest {
 
         TagAssertion[] tags = featureDetails.getTags();
         assertThat(tags).hasSize(1);
-        tags[0].getLink().hasLabelAndAddress("@featureTag", "report-tag_featureTag.html");
+        tags[0].getLink().hasLabelAndAddress("@featureTag", "report-tag_2956005635.html");
 
         assertThat(featureDetails.getDescription()).isEqualTo(feature.getDescription());
     }
